Relocate missed BQ IT tests
