+++ scio-google-cloud-platform/src/it/scala/com/spotify/scio/bigquery/TypedBigQueryIT.scala
